chore(deps): replace dependency on `lazy_static` by `once_cell`
`once_cell` is a more ergonomic alternative to `lazy_static` which
provides a more straightforward syntax for lazily initialized statics.
Its API is also closer to the nightly `std::sync::LazyLock`, which
should make it easier to migrate to it in the event that it gets
stabilized. `once_cell` is also pulled in transitively more often than
`lazy_static` by the dependencies of this project, and I think it's
likely that the ecosystem will move to `once_cell`.