fix(lexer/v2-full): fix injecting pp. directives after ext. declarations
When using the v2-full lexer it is necessary to inject at least some
preprocessor directives back to the AST so they appear in the transpiled
GLSL.

The lexer supports this use case for preprocessor directives in external
declaration position, but the tests that cover this injection
functionality do not check that the results are correct when there are
several external declarations in a translation unit.

Given that the order of preprocessor directives is significant, this is
an incorrect behavior. I pinned down the cause of this problem to a bad
computation of the declaration start position in the injection code.

These changes improve the AST directive injection in this case and add a
test for this problem.