feat(transpiler): overhaul formatting and customizability
While the current GLSL transpiler output format is quite customizable, I
needed to further extend its customizability to better support outputting
GLSL in both pretty and minified formats. To that end, these changes
overhaul the transpiler's formatting logic and implement some minor
format changes and fixes. The fixes included are complementary to those in
https://github.com/vtavernier/glsl-lang/pull/19.

The full list of changes, in no particular order, is as follows:

- Added a feature to collapse compound statements with a single statement
to the statement itself, but only if the compound statement is not part of a
function definition, in which case case the GLSL grammar requires a compound
statement. This collapsing feature is toggleable and disabled by default.

- Added formatting options to toggle whether spaces are inserted around
open block braces, after case labels, before the `else` keyword, after the
list separator and for separator tokens (comma and semicolon, respectively),
around the initializer list expression braces, and around the parentheses that
are part of some statements, such as selection, jump, or switch statements.

- Replaced instances of `write_str` of a single char with `write_char`.

- Removed a extraneous whitespace before the opening parenthesis of a layout
specifier.

- Expanded the set of operator-like tokens written with `write_binary_op` so
that the value of the corresponding format option is applied more consistently.

- Fixed not outputting the semicolon required to terminate the expression
statement in a for init statement.

- Removed excess whitespace after the `return` keyword when no expression
follows.

- Updated tests.