fix(transpiler): put pp. dirs on their own line no matter what the terminators are
Currently, when the formatting settings instruct the transpiler to use
terminators that do not create newlines, it is possible for preprocessor
directives to share a line with other external declarations, which
violates the GLSL specification.

For example, given the following input:

```glsl
 #version 110
vec3 u;
vec3 v;
 #pragma "hi"
vec3 w;
```

The current transpiler would generate the following invalid output when
`declaration_terminator` is set to `Whitespace::None`:

```glsl
 #version 110
vec3 u;vec3 v;#pragma "hi"
vec3 w;
```

To fix this issue, let's make the formatting state capable of making
sure that the output cursor is at an empty line for preprocessor
directives, and make sure that preprocessor directives force a line
flush to occur for the next external declaration.

While at it, I've fixed the else-if preprocessor directive incorrectly
being transpiled to #elseif instead of #elif.

These changes depend on https://github.com/vtavernier/glsl-lang/pull/15
for the tests to pass with the v2-full lexer, as they expect such
directives to be injected at the expected positions.