fix(transpiler): add missing brackets for array expressions
As defined in the GLSL grammar, postfix expressions may be suffixed by
an integer expression between parentheses to index primary expressions
like matrices and arrays. The AST represents such suffix with a Bracket
node.

This node was slightly refactored while being ported from the glsl crate
to contain an expression instead of an array specifier, which aligns
better with the GLSL grammar, but introduced a regression due to
omitting an explicit write of the necessary brackets. For example, this
input:

```glsl
void main() { ProjMat[0][0] = 0.0; }
```

Would transpile to the following incorrect output:

```glsl
void main() { ProjMat00 = 0.0; }
```

To fix this, surround the integer expression between brackets as
mandated by the GLSL grammar and add a test for this.